add node to node examples
